Move API into a dedicated folder
